+++ client/src/components/messages/Message.js
@@ -112,21 +112,6 @@ function Message() {
   }, [chatUserData]);
 
   useEffect(() => {
-    //join room for this conversation
-    if (chatUserData) {
-      const data = {
-        userId: user._id,
-        chatId: chatUserData.chatId,
-      };
-      socket && socket.emit("join", { data }, () => {});
-    }
-    //clean socket before destroying component
-    return () => {
-      socket && socket.off();
-    };
-  }, [chatUserData, socket]);
-
-  useEffect(() => {
     socket &&
       socket.on("newMessage", data => {
         const newMessage = {
@@ -139,7 +124,7 @@ function Message() {
         lastMessageRef.current = null;
         setMessage(newMessage);
       });
-  }, [chatUserData, socket]);
+  }, [socket]);
 
   useEffect(() => {
     if (
@@ -173,6 +158,7 @@ function Message() {
       timeCreated: moment().format(),
       chatId: chatUserData.chatId,
       wasRead: false,
+      room: chatUserData.userId,
     });
 
     e.preventDefault();
@@ -180,6 +166,8 @@ function Message() {
       content: messageContent,
       chatId: chatUserData.chatId,
     });
+    setMessageContent("");
+    getMessageHistory();
   };
 
   return (
